Feature: Store state in World instead of Creature
Previously, historical state tracking was done through the individual
actors in the simulation. This is a big problem when trying to rebuild
the state of the simulation at any given time.

Now, state is tracked through a vector of WorldState structs. This makes
it much easier to replay the simulation.

Package: creature-simulation@0.1.0
Contact: Sam Christy <samchristywork@gmail.com>