Bug fix: Fixed bounded movement
Previously, the places that the creatures moved to when they encountered
the end of the world made no sense. This was due to a bug in the logic
of the bounding code. I also restricted the map size so that it takes a
constant amount of space as long as the terminal is large enough to
handle it. This change was made for those same bounding reasons.

Package: creature-simulation@0.1.0
Contact: Sam Christy <samchristywork@gmail.com>