Optimization: Plants are no longer real objects
Plants used to be objects that were instantiated using `new`. This meant
that in order to detect that it was at the same position as a plant a
creature needed to loop through every plant in the array. This was
pretty inefficient, and is much less flexible than the new method that I
have in mind.

Package: creature-simulation@0.1.0
Contact: Sam Christy <samchristywork@gmail.com>